% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{abstract}
\alias{abstract}
\title{get nsfc abstract}
\usage{
abstract(url, header, subject, search, yearStart, yearEnd,
  year_ascend = TRUE, itemCategory, fundStart, fundEnd)
}
\arguments{
\item{url}{url}

\item{header}{header}

\item{subject}{subject, ex:H0801}

\item{search}{string to search, ex: "m6A"}

\item{yearStart}{number, year of start}

\item{yearEnd}{number, year of start}

\item{year_ascend}{a logical argument, default is TRUE, ordre by year ascending}

\item{itemCategory}{category}

\item{fundStart}{fund min}

\item{fundEnd}{fund max}
}
\value{
dataframe
}
\description{
get nsfc abstract from http://www.sciencenet.cn/, by key
}
\examples{
abstract(yearStart=2018)
}
