% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsfc.R
\name{nsfc}
\alias{nsfc}
\title{get nsfc items}
\usage{
nsfc(url, headers, subject, yearStart, yearEnd, itemCategory, fundStart,
  fundEnd, search)
}
\arguments{
\item{url}{url}

\item{headers}{headers}

\item{subject}{subject, ex:H0801}

\item{yearStart}{number, year of start}

\item{yearEnd}{number, year of start}

\item{itemCategory}{category}

\item{fundStart}{fund min}

\item{fundEnd}{fund max}

\item{search}{string to search}
}
\value{
dataframe with 7 items
}
\description{
get nsfc items
}
\examples{
nsfc(yearStart=2018)
}
