% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\title{get nsfc subject}
\usage{
subject(url, header, subject, yearStart, yearEnd, year_ascend = TRUE)
}
\arguments{
\item{url}{url}

\item{header}{header}

\item{subject}{subject, ex:H0801}

\item{yearStart}{number, year of start}

\item{yearEnd}{number, year of start}

\item{year_ascend}{a logical argument, default is TRUE, ordre by year ascending}
}
\value{
dataframe
}
\description{
get nsfc subject from http://www.sciencenet.cn/, by key
}
\examples{
subject(yearStart=2018)
}
